Separate GJK test cases
